Save all notes and references in workshop js file and create new main.js file with code to be tested
